requirements.txt
streamlit==1.27.2
Pillow==10.0.1
openai==0.28.1